% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbaser-package.R
\docType{package}
\name{brainbaser-package}
\alias{brainbaser-package}
\alias{brainbaser}
\title{Programmatic Access to the Brainbase Drosophila Neuroanatomy Website}
\description{
This package provides programmatic access to some functions of the Brainbase
Drosophila neuroanatomy website, \url{brainbase.imp.ac.at}. In the first
version of the package functionality is focussed on overlap scores of
expression patterns for different neuropil domains.
}
\references{
The Vienna Tiles (VT) driver collection (but not its expression
patterns in the adult brain) has been published in:

Genome-scale functional characterization of Drosophila developmental
enhancers in vivo. Kvon EZ, Kazmar T, Stampfel G, Yanez-Cuna JO, Pagani M,
Schernhuber K, Dickson BJ, Stark A. Nature. 2014 Aug 7;512(7512):91-5
\href{http://dx.doi.org/10.1038/nature13395}{doi:10.1038/nature13395}
}
\seealso{
\code{\link{neuropil_overlaps}}, \code{\link{neuropils}}
}
\keyword{package}
