% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuropil_overlaps.R
\name{neuropil_overlaps}
\alias{neuropil_overlaps}
\title{Fetch fractional overlap for each neuropil for all images in BrainBase}
\usage{
neuropil_overlaps(nps, raw = FALSE)
}
\arguments{
\item{nps}{The neuropils to fetch as integer ids. A non-finite value implies
all neuropils.}

\item{raw}{Whether to return raw JSON result}
}
\value{
processed data.frame
}
\description{
Fetch fractional overlap for each neuropil for all images in BrainBase
}
\examples{
x=neuropil_overlaps(1152)
}

