% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neuropils}
\alias{neuropils}
\title{Data frame containing definitions of standard Brainbase neuropil domains}
\description{
These were retrieved by carrying out a query for all neuropil regions
}

