% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuropil_overlaps.R
\name{neuropil_overlaps}
\alias{neuropil_overlaps}
\title{Fetch fractional overlap for each neuropil for all images in BrainBase}
\usage{
neuropil_overlaps(nps, raw = FALSE, normalised_by = c("neuropil",
  "staining"))
}
\arguments{
\item{nps}{The neuropils to fetch, specified either as integer ids or as the
character vector abbreviated neuropil name (see \code{\link{neuropils} in
both cases}. The special values of "brain" and "vnc" imply all brain or vnc
neuropils, respectively.}

\item{raw}{Whether to return the raw httr result}

\item{normalised_by}{How to normalise the results. This denominator can
  either be the total number of voxels in each \bold{neuropil} region or the
  total amount of \code{staining} in the whole expression pattern.}
}
\value{
processed \code{data.frame} with columns \itemize{

  \item imgId Integer image id

  \item name Image name

  \item ... additional columns expressing the fraction of each neuropil
  domain occupied by foreground voxels and titled with the abbreviated
  neuropil names.}
}
\description{
Fetch fractional overlap for each neuropil for all images in BrainBase
}
\examples{
x=neuropil_overlaps(1152)

\dontrun{
r=neuropil_overlaps(1152, raw=TRUE)
json_content=content(r)

all_brain=neuropil_overlaps('brain')
all_vnc=neuropil_overlaps('vnc')
}
}
\seealso{
The \code{\link{neuropils}} data.frame contains details of the
  different neuropil regions.
}

