% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuropil_overlaps.R
\name{neuropil_overlaps}
\alias{neuropil_overlaps}
\title{Fetch fractional overlap for each neuropil for all images in BrainBase}
\usage{
neuropil_overlaps(nps, raw = FALSE)
}
\arguments{
\item{nps}{The neuropils to fetch, specified either as integer ids or as the
character vector abbreviated neuropil name (see \code{\link{neuropils} in
both cases}. The special values of "brain" and "vnc" imply all brain or vnc
neuropils, respectively.}

\item{raw}{Whether to return raw JSON result}
}
\value{
processed data.frame
}
\description{
Fetch fractional overlap for each neuropil for all images in BrainBase
}
\examples{
x=neuropil_overlaps(1152)
}
\seealso{
The \code{\link{neuropils}} data.frame contains details of the
  different neuropil regions.
}

